clean code, derive weights symbolically, add docs
