+++ setup.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+set +e
+virtualenv env
+source env/bin/activate
+pip install -r requirements.txt
+echo
+echo "Done."
